fix tests in js
