--- spec/juno-spec.d.ts
@@ -1,1 +1,0 @@
-export {};
