--- spec/runner.d.ts
@@ -1,2 +1,0 @@
-declare const _default: any;
-export default _default;
