--- spec/tsconfig.tsbuildinfo
@@ -1,3 +1,0 @@
-{
-  "version": "3.8.3"
-}
