--- spec_src/runner.ts
@@ -1,8 +1,0 @@
-import { createRunner } from "atom-jasmine3-test-runner"
-
-// https://github.com/UziTech/atom-jasmine3-test-runner#api
-export default createRunner({
-  timeReporter : true,
-  specHelper: true,
-  testPackages: ["language-julia", "ink"]
-})
