--- spec_src/tsconfig.json
@@ -1,37 +1,0 @@
-{
-  "compilerOptions": {
-    //// Linting Options - Uncomment options to get more features (usually more restrictive)
-    // "strict": true,    // includes all of the following and more
-    "strictNullChecks": true,
-    // "forceConsistentCasingInFileNames": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noImplicitAny": true,
-    // "noImplicitThis": true,
-    "noFallthroughCasesInSwitch": true,
-    "allowJs": true,
-    //// Compilation options
-    "declaration": true,
-    "emitDecoratorMetadata": true,
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "incremental": true,
-    // "inlineSourceMap": false,
-    // "preserveConstEnums": true,
-    "sourceMap": true,
-    "preserveSymlinks": true,
-    // "removeComments": true,
-    // "jsx": "react",
-    // "jsxFactory": "etch.dom",
-    "lib": ["ES2018", "dom"],
-    "target": "ES2018",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    // "noLib": false,
-    // "importHelpers": true, // if true you should add tslib to deps
-    // "skipLibCheck": false,
-     "outDir": "../spec"
-  },
-  "compileOnSave": true
-}
